// apis/database/types/config.ts
export interface ConfigFromServer {
  defaultPreset?: unknown;
  maxComparedWheels?: unknown;
  maxCurrentAllowed?: unknown;
  paginationSize?: unknown;

  calculatedRange?: unknown;
  prices?: unknown;
  purchaseLinks?: unknown;

  specColumns?: unknown;
  
  listAdditionalSpecs?: unknown;
  listMainSpecs?: unknown;
  
  wheelAdditionalSpecs?: unknown;
  wheelHighlightedSpecs?: unknown;
  wheelMainSpecs?: unknown;
}